refactor: move the bowtie2 mapping index into the workflow runtime
`createMappingIndex`, `buildBowtie2Index`, `getBowtie2BuildVersion` and
`buildMappingIndexCacheParams` were pathoscope's, but the two namespaces they
derive are shared with Python and between every workflow that maps against a
reference or a subtraction. Apps cannot import each other, so nuvs could not
reach them.

`workflow` becomes a parameter rather than a hardcoded "pathoscope": it is a
field in every derived key, so each workflow shares with its own Python
counterpart and not with the other. The version-string matcher splits out to
`subprocess/version.ts`, which is what `getCdHitEstVersion` still needs.

The byte-exact Python goldens stay in the app, next to the `extra` params
they pin, and are unchanged — the derivation did not move. The runtime's own
test covers what is generic, plus that two workflow names derive different
keys.
