+++ packages/workflow/src/index.ts
@@ -15,10 +15,12 @@ export * from "./index/queries";
 export * from "./index/schema";
 export * from "./lifecycle/claim";
 export * from "./lifecycle/ping";
+export * from "./mapping/mappingIndex";
 export * from "./run";
 export * from "./serializable";
 export * from "./step";
 export * from "./subprocess/execa";
 export * from "./subprocess/lines";
 export * from "./subprocess/types";
+export * from "./subprocess/version";
 export * from "./workPath";
